backend/requirements.txt
fastapi>=0.112
uvicorn[standard]>=0.29
pydantic>=2.7
google-generativeai>=0.7
groq>=0.9

